+++ tools/luban-provisioner/templates/source/luban-dagster-dbt-starrocks-code-location-source-template/{{cookiecutter.app_name}}/src/{{cookiecutter.package_name}}/assets/dbt.py
@@ -20,7 +20,7 @@ dbt_project = DbtProject(
     packaged_project_dir=dbt_project_dir,
 )
 
-prepare_manifest_if_missing(target=dbt_target)
+prepare_manifest_if_missing()
 
 
 @dbt_assets(